@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^20.3.2",
-    "@angular/cdk": "^20.2.5",
-    "@angular/common": "^20.3.2",
-    "@angular/compiler": "^20.3.2",
-    "@angular/core": "^20.3.2",
-    "@angular/forms": "^20.3.2",
-    "@angular/platform-browser": "^20.3.2",
-    "@angular/platform-browser-dynamic": "^20.3.2",
-    "@angular/router": "^20.3.2",
+    "@angular/animations": "^21.0.5",
+    "@angular/cdk": "^21.0.3",
+    "@angular/common": "^21.0.5",
+    "@angular/compiler": "^21.0.5",
+    "@angular/core": "^21.0.5",
+    "@angular/forms": "^21.0.5",
+    "@angular/platform-browser": "^21.0.5",
+    "@angular/platform-browser-dynamic": "^21.0.5",
+    "@angular/router": "^21.0.5",
     "@ngx-translate/core": "^17.0.0",
     "@primeng/themes": "^20.2.0",
     "chart.js": "^4.5.0",
@@ -36,9 +36,9 @@
     "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^20.3.3",
-    "@angular/cli": "^20.3.3",
-    "@angular/compiler-cli": "^20.3.2",
+    "@angular-devkit/build-angular": "^21.0.3",
+    "@angular/cli": "^21.0.3",
+    "@angular/compiler-cli": "^21.0.5",
     "@types/jasmine": "~5.1.9",
     "autoprefixer": "^10.4.21",
     "eslint": "^9.36.0",
@@ -55,6 +55,6 @@
     "postcss": "^8.4.49",
     "prettier": "^3.6.2",
     "tailwindcss": "^3.4.17",
-    "typescript": "~5.8.3"
+    "typescript": "~5.9.3"
   }
 }
