@@ -19,18 +19,15 @@ interface Breadcrumb {
     standalone: true,
     imports: [CommonModule, RouterModule],
     template: `<nav class="layout-breadcrumb">
-        <ol>
-            <ng-template
-                ngFor
-                let-item
-                let-last="last"
-                [ngForOf]="breadcrumbs$ | async"
-            >
-                <li>{{ item.label }}</li>
-                <li *ngIf="!last" class="layout-breadcrumb-chevron">/</li>
-            </ng-template>
-        </ol>
-    </nav> `,
+          <ol>
+            @for (item of breadcrumbs$ | async; track item; let last = $last) {
+              <li>{{ item.label }}</li>
+              @if (!last) {
+                <li class="layout-breadcrumb-chevron">/</li>
+              }
+            }
+          </ol>
+        </nav>`,
 })
 export class AppBreadcrumb {
     private readonly _breadcrumbs$ = new BehaviorSubject<Breadcrumb[]>([]);
