@@ -1,4 +1,4 @@
-import { CommonModule } from '@angular/common';
+
 import { Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { MenuItem } from 'primeng/api';
@@ -8,19 +8,22 @@ import { AppMenuitem } from '../menuitem/menuitem.component';
 @Component({
     selector: 'app-menu',
     standalone: true,
-    imports: [CommonModule, AppMenuitem, RouterModule],
+    imports: [AppMenuitem, RouterModule],
     template: `<ul class="layout-menu">
-        <ng-container *ngFor="let item of model; let i = index">
-            <li
+          @for (item of model; track item; let i = $index) {
+            @if (!item.separator) {
+              <li
                 app-menuitem
-                *ngIf="!item.separator"
                 [item]="item"
                 [index]="i"
                 [root]="true"
-            ></li>
-            <li *ngIf="item.separator" class="menu-separator"></li>
-        </ng-container>
-    </ul> `,
+              ></li>
+            }
+            @if (item.separator) {
+              <li class="menu-separator"></li>
+            }
+          }
+        </ul>`,
 })
 export class AppMenu {
     model: MenuItem[] = [];
