@@ -15,32 +15,37 @@ import {LayoutService} from '@/layout/service/layout.service';
     imports: [CommonModule, RouterModule, RippleModule, TooltipModule],
     template: `
         <ng-container>
-            <div *ngIf="root && item.visible !== false" class="layout-menuitem-root-text">
-                {{ item.label }}
+          @if (root && item.visible !== false) {
+            <div class="layout-menuitem-root-text">
+              {{ item.label }}
             </div>
+          }
+          @if ((!item.routerLink || item.items) && item.visible !== false) {
             <a
-                *ngIf="(!item.routerLink || item.items) && item.visible !== false"
-                [attr.href]="item.url"
-                (click)="itemClick($event)"
-                (mouseenter)="onMouseEnter()"
-                [ngClass]="item.class"
-                [attr.target]="item.target"
-                tabindex="0"
-                pRipple
-                [pTooltip]="item.label"
-                [tooltipDisabled]="!(isSlim() && root && !active)"
-            >
-                <i [ngClass]="item.icon" class="layout-menuitem-icon"></i>
-                <span class="layout-menuitem-text">{{ item.label }}</span>
-                <i class="pi pi-fw pi-angle-down layout-submenu-toggler" *ngIf="item.items"></i>
+              [attr.href]="item.url"
+              (click)="itemClick($event)"
+              (mouseenter)="onMouseEnter()"
+              [ngClass]="item.class"
+              [attr.target]="item.target"
+              tabindex="0"
+              pRipple
+              [pTooltip]="item.label"
+              [tooltipDisabled]="!(isSlim() && root && !active)"
+              >
+              <i [ngClass]="item.icon" class="layout-menuitem-icon"></i>
+              <span class="layout-menuitem-text">{{ item.label }}</span>
+              @if (item.items) {
+                <i class="pi pi-fw pi-angle-down layout-submenu-toggler"></i>
+              }
             </a>
+          }
+          @if (item.routerLink && !item.items && item.visible !== false) {
             <a
-                *ngIf="item.routerLink && !item.items && item.visible !== false"
-                (click)="itemClick($event)"
-                (mouseenter)="onMouseEnter()"
-                [ngClass]="item.class"
-                [routerLink]="item.routerLink"
-                routerLinkActive="active-route"
+              (click)="itemClick($event)"
+              (mouseenter)="onMouseEnter()"
+              [ngClass]="item.class"
+              [routerLink]="item.routerLink"
+              routerLinkActive="active-route"
                 [routerLinkActiveOptions]="
                     item.routerLinkActiveOptions || {
                         paths: 'exact',
@@ -49,31 +54,36 @@ import {LayoutService} from '@/layout/service/layout.service';
                         fragment: 'ignored'
                     }
                 "
-                [fragment]="item.fragment"
-                [queryParamsHandling]="item.queryParamsHandling"
-                [preserveFragment]="item.preserveFragment"
-                [skipLocationChange]="item.skipLocationChange"
-                [replaceUrl]="item.replaceUrl"
-                [state]="item.state"
-                [queryParams]="item.queryParams"
-                [attr.target]="item.target"
-                tabindex="0"
-                pRipple
-                [pTooltip]="item.label"
-                [tooltipDisabled]="!(isSlim() && root)"
+            [fragment]="item.fragment"
+            [queryParamsHandling]="item.queryParamsHandling"
+            [preserveFragment]="item.preserveFragment"
+            [skipLocationChange]="item.skipLocationChange"
+            [replaceUrl]="item.replaceUrl"
+            [state]="item.state"
+            [queryParams]="item.queryParams"
+            [attr.target]="item.target"
+            tabindex="0"
+            pRipple
+            [pTooltip]="item.label"
+            [tooltipDisabled]="!(isSlim() && root)"
             >
-                <i [ngClass]="item.icon" class="layout-menuitem-icon"></i>
-                <span class="layout-menuitem-text">{{ item.label }}</span>
-                <i class="pi pi-fw pi-angle-down layout-submenu-toggler" *ngIf="item.items"></i>
-            </a>
-
-            <ul #submenu *ngIf="item.items && item.visible !== false" [@children]="submenuAnimation" (@children.done)="onSubmenuAnimated($event)">
-                <ng-template ngFor let-child let-i="index" [ngForOf]="item.items">
-                    <li app-menuitem [item]="child" [index]="i" [parentKey]="key" [class]="child['badgeClass']"></li>
-                </ng-template>
-            </ul>
+            <i [ngClass]="item.icon" class="layout-menuitem-icon"></i>
+            <span class="layout-menuitem-text">{{ item.label }}</span>
+            @if (item.items) {
+              <i class="pi pi-fw pi-angle-down layout-submenu-toggler"></i>
+            }
+          </a>
+        }
+        
+        @if (item.items && item.visible !== false) {
+          <ul #submenu [@children]="submenuAnimation" (@children.done)="onSubmenuAnimated($event)">
+            @for (child of item.items; track child; let i = $index) {
+              <li app-menuitem [item]="child" [index]="i" [parentKey]="key" [class]="child['badgeClass']"></li>
+            }
+          </ul>
+        }
         </ng-container>
-    `,
+        `,
     animations: [
         trigger('children', [
             state(
