@@ -1,4 +1,4 @@
-import { CommonModule } from '@angular/common';
+
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
@@ -24,7 +24,6 @@ import { ContoBancarioListingEliminaMessaggioPipe } from '../pipes/conto-bancari
 @Component({
   selector: 'app-conto-bancario-listing',
   imports: [
-    CommonModule,
     FormsModule,
     CardModule,
     ButtonModule,
@@ -35,7 +34,7 @@ import { ContoBancarioListingEliminaMessaggioPipe } from '../pipes/conto-bancari
     SkeletonModule,
     TooltipModule,
     ContoBancarioListingEliminaMessaggioPipe
-  ],
+],
   providers: [
     ContoBancarioService
   ],
