@@ -1,4 +1,4 @@
-import { CommonModule } from '@angular/common';
+
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
@@ -18,13 +18,12 @@ import { FormInputTextareaComponent } from 'src/shared/form-input-textarea/form-
 @Component({
   selector: 'app-stampante-manager',
   imports: [
-    CommonModule,
     FormsModule,
     FormInputTextComponent,
     FormInputTextareaComponent,
     CardModule,
     ButtonModule
-  ],
+],
   providers: [
     StampanteService
   ],
